Switched to XCode 4.2 and Adium 1.5!

+++ AISmileyController.m
@@ -59,7 +59,7 @@ static NSMutableArray* smileyArray = nil;
 	return cache_path;
 }
 
-int compareSmiley(id first, id second, void* context)
+NSInteger compareSmiley(id first, id second, void* context)
 {
 	NSString* first_shortcut=[first shortcut];
 	NSString* second_shortcut=[second shortcut];
